+++ Source/Global/global.h
@@ -1,9 +1,6 @@
 // Copyright (c) Microsoft Corporation
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 #pragma once
-#include "win/utils_win.h"
-#include "taskImpl.h"
 
 NAMESPACE_XBOX_HTTP_CLIENT_BEGIN
