+++ Source/Global/mem.cpp
