+++ Source/Mock/mock.cpp
@@ -2,8 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 #include "pch.h"
-#include "http_mock.h"
-#include "../httpcall.h"
+#include "mock.h"
+#include "../http/httpcall.h"
 
 using namespace xbox::httpclient;
